Updated rendering pipeline, added light support.
Changed the naive rendering technique for a more intelligent one, saving
a lot of overhead. Added diffuse, and specular lighting support in the
shader programs, by using the camera vector, the vector pointing from
the object towards the light, and the object's surface normal vectors,
as well as a specular factor (which is a dot product between the
reflected light direction, and the unit vector towards the camera. The
more "alike" they are, the closer to 1.0 their dot product will be).
This specular factor is raised to the power of the shine damper, which
is a property of the Light object, passed as an uniform variable to the
fragment shader.